+++ setup.cfg
@@ -43,6 +43,7 @@ asyncio =
     greenlet!=0.4.17
 mypy =
     mypy >= 0.910
+    types-greenlet >= 2
 mssql = pyodbc
 mssql_pymssql = pymssql
 mssql_pyodbc = pyodbc
@@ -79,7 +80,6 @@ asyncmy =
 aiosqlite =
     %(asyncio)s
     aiosqlite
-    typing_extensions!=3.10.0.1
 sqlcipher =
     sqlcipher3_binary
