+++ tox.ini
@@ -176,6 +176,7 @@ commands=
 deps=
      greenlet != 0.4.17
      mypy >= 1.6.0
+     types-greenlet
 commands =
     mypy  {env:MYPY_COLOR} ./lib/sqlalchemy
     # pyright changes too often with not-exactly-correct errors
@@ -189,7 +190,7 @@ deps=
      greenlet != 0.4.17
      mypy >= 1.2.0
      patch==1.*
-
+     types-greenlet
 commands =
     pytest {env:PYTEST_COLOR} -m mypy {posargs}
